BUG fix: too many files opened.
Very important bugs fixed: too many files opened. It was caused by heavy use of recursion, and making a lot of files and folders stay open.

Now we use a global array to keep track of folders that are open. And use depth-first search to traverse through all the folders to search .c or .h files.

Note: in the future, file extensions may be parameters so that we can check other files as well like .cpp, .py, or make files etc.